+++ libc/mem.h
@@ -1,7 +1,11 @@
 #ifndef MEM_H
 #define MEM_H
 
+#include <stdbool.h>
+#include <stdint.h>
+#include <stddef.h>
 
-void memcpy(void* dst, const void* src, int nbytes);
+
+void memcpy(void* dst, const void* src, size_t nbytes);
 
 #endif
